requirements.txt
click==8.0.1
Flask==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
Werkzeug==2.0.1
requests==2.26.0
grpcio==1.39.0
sqlalchemy==1.4
python-dateutil==2.8.2
jsonschema==3.2.0
googlemaps==4.5.3
Pillow==8.3.2
mysql-connector-python
mysqlclient==2.0.3
